+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## [0.2.6](https://github.com/fxwiegand/predictosaurus/compare/v0.2.5...v0.2.6) (2024-12-16)
+
+
+### Bug Fixes
+
+* Create output dir if non existent ([79b1383](https://github.com/fxwiegand/predictosaurus/commit/79b1383075071d506bbfbeff13e2622a5295dd54))
+* Fix feature graph generation ([283feec](https://github.com/fxwiegand/predictosaurus/commit/283feec842e031a288dacbdbd1e28fd0a4ebfb5d))
+
+## [0.2.5](https://github.com/fxwiegand/predictosaurus/compare/v0.2.4...v0.2.5) (2024-12-12)
+
+
+### Bug Fixes
+
+* Remove primary key constraint on nodes table ([e62ac37](https://github.com/fxwiegand/predictosaurus/commit/e62ac3759bfbc9ae410b6101b82f53485d28be1b))
+
+## [0.2.4](https://github.com/fxwiegand/predictosaurus/compare/v0.2.3...v0.2.4) (2024-12-12)
+
+
+### Bug Fixes
+
+* Fix wrong interpretation of probabilities ([7aeadb4](https://github.com/fxwiegand/predictosaurus/commit/7aeadb4a81462c6b1afd8ec8fba5506d39ac8f23))
+
+
+### Performance Improvements
+
+* Add indexes for write_graphs tables ([731ea91](https://github.com/fxwiegand/predictosaurus/commit/731ea91120e6ea8aee835aa0a01e02207c566cbf))
+
 ## [0.2.3](https://github.com/fxwiegand/predictosaurus/compare/v0.2.2...v0.2.3) (2024-12-10)